requirements.txt
altgraph==0.17
app==0.0.1
appdirs==1.4.3
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==3.0.3
django-chartit==0.2.9
django-chartit2==0.2.2
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-excel==0.0.10
django-filter==2.2.0
django-grappelli==2.13.3
django-import-export==2.0.2
djangorestframework==3.11.0
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
geojson==2.5.0
idna==2.8
jdcal==1.4.1
kiwisolver==1.1.0
lml==0.0.9
Markdown==3.2.2
MarkupPy==1.14
matplotlib==3.1.3
mysql-client==0.0.1
numpy==1.18.1
odfpy==1.4.1
openpyxl==3.0.3
pefile==2019.4.18
Pillow==7.0.0
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
PyInstaller==3.6
PyMySQL==0.9.3
pyowm==2.10.0
pyparsing==2.4.6
pyTelegramBotAPI==3.6.7
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.22.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
tablib==1.1.0
texttable==1.6.2
urllib3==1.25.8
virtualenv==20.0.4
xlrd==1.2.0
xlwt==1.3.0
